add a logout dialog to notify users

+++ lib/pages/profile_page.dart
@@ -190,13 +190,44 @@ class _ProfilePageState extends State<ProfilePage> {
                     const SizedBox(height: 20),
                     ElevatedButton.icon(
                       onPressed: () async {
-                        // clean user data in local storage
-                        SharedPreferences prefs =
-                            await SharedPreferences.getInstance();
-                        await prefs.clear();
+                        // Show confirmation dialog
+                        bool confirmLogout = await showDialog(
+                          context: context,
+                          builder: (context) {
+                            return AlertDialog(
+                              title: const Text("Confirm Logout"),
+                              content: const Text(
+                                  "Are you sure you want to log out?"),
+                              actions: [
+                                TextButton(
+                                  onPressed: () {
+                                    Navigator.of(context)
+                                        .pop(false); // Cancel logout
+                                  },
+                                  child: const Text("Cancel"),
+                                ),
+                                TextButton(
+                                  onPressed: () {
+                                    Navigator.of(context)
+                                        .pop(true); // Confirm logout
+                                  },
+                                  child: const Text("Logout"),
+                                ),
+                              ],
+                            );
+                          },
+                        );
 
-                        // nav to login page
-                        Navigator.of(context).pushReplacementNamed('/login');
+                        // If user confirmed, proceed with logout
+                        if (confirmLogout == true) {
+                          // Clean user data in local storage
+                          SharedPreferences prefs =
+                              await SharedPreferences.getInstance();
+                          await prefs.clear();
+
+                          // Navigate to login page
+                          Navigator.of(context).pushReplacementNamed('/login');
+                        }
                       },
                       icon: const Icon(
                         Icons.logout,
@@ -212,7 +243,7 @@ class _ProfilePageState extends State<ProfilePage> {
                           borderRadius: BorderRadius.circular(30),
                         ),
                       ),
-                    ),
+                    )
                   ],
                 );
               }
